requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
folium>=0.14.0
pandas>=2.0.0
